+++ pcs/cli/common/console_report.py
@@ -357,6 +357,11 @@ CODE_TO_MESSAGE_BUILDER_MAP = {
         .format(**info)
     ,
 
+    codes.NODE_COMMUNICATION_ERROR_TIMED_OUT: lambda info:
+        "{node}: Connection timeout ({reason})"
+        .format(**info)
+    ,
+
     codes.COROSYNC_CONFIG_DISTRIBUTION_STARTED:
         "Sending updated corosync.conf to nodes..."
     ,