+++ pcs/settings_default.py
@@ -44,3 +44,4 @@ sbd_config = "/etc/sysconfig/sbd"
 pacemaker_wait_timeout_status = 62
 booth_config_dir = "/etc/booth"
 booth_binary = "/usr/sbin/booth"
+default_request_timeout = 60